BACKWARDS INCOMPATIBLE: Hash instances by object id rather than value
This patch updates our usage of `attr.s` to set `hash=False`, so that we
use id-based hashing instead of value-based hashing.

attrs 17.1.0 changed the default for `hash` to `None`, which makes
objects unhashable.  We set `hash=False` so that we can continue to use
objects as keys in dictionaries, but without attempting to hash by
value.

http://www.attrs.org/en/stable/changelog.html